reorganize all tests for pantable.py into subfolder
