--- tests/context.py
@@ -1,5 +1,0 @@
-import os
-import sys
-sys.path.insert(0, os.path.abspath('..'))
-
-from pantable.pantable import *
