+++ tests/pantable/context.py
@@ -1,0 +1,5 @@
+import os
+import sys
+sys.path.insert(0, os.path.abspath('..' + os.sep + '..'))
+
+from pantable.pantable import *
