+++ tests/pantable/test_init_table_options.py
