+++ tests/pantable/test_parse_table_list.py
