+++ tests/pantable/test_parse_table_options.py
