+++ tests/pantable/test_regularize_table_list.py
