+++ tests/pantable/test_to_bool.py
