+++ tests/test_pantable.md
@@ -107,7 +107,7 @@ caption: "*Great* Title"
 header: True
 markdown: True
 alignment: AlignLeft, AlignRight, AlignCenter, AlignDefault
-include: tests/csv_tables.csv
+include: tests/pantable/csv_tables.csv
 ...
 ```
